requirements.txt
streamlit
yfinance
pandas
pandas-ta
supabase
google-generativeai
curl_cffi